Settings: Fix stray divider in connected devices screen
When no Bluetooth devices are nearby or no slice is present, there is a
stray preference divider at the top of screen that makes the action bar's
shadow look inconsistent. This removes the divider below the slice,
which is likely not that important for visuals anyway, to fix the action
bar appearance issue.

Change-Id: I6c5d0d00fcf7734b8980cc8c50c248aaf105b25f